Fix the copyright symbol encoding
In some files the copyright symbol is encoded in ISO-8859-1 as single
byte (0xA9) which is incompatible with UTF-8. This causes most text
editors and IDEs to render the symbol as � because by default they
assume a file is encoded in UTF-8 when no byte order marks are present.

Also fixes apostrophe symbols in HtmlWeb.cs file.